fix(testapp): do not render the application shell in activator microfrontends
