+++ apps/microfrontend-platform-testing-app/src/app/app-routing.module.ts
@@ -23,26 +23,33 @@ import { PreferredSizeComponent } from './preferred-size/preferred-size.componen
 import { PlatformPropertiesComponent } from './platform-properties/platform-properties.component';
 import { LookupCapabilityComponent } from './manifest/lookup-capability/lookup-capability.component';
 import { LookupIntentionComponent } from './manifest/lookup-intention/lookup-intention.component';
+import { AppShellComponent } from './app-shell/app-shell.component';
 
 const routes: Routes = [
-  {path: 'browser-outlets', component: BrowserOutletsComponent, data: {pageTitle: 'Allows displaying web content in one or more browser outlets', matrixParams: new Map().set('count', 2), pageTitleVisible: false}},
-  {path: 'router-outlet', component: RouterOutletComponent, data: {pageTitle: 'Allows displaying web content in a router outlet'}},
-  {path: 'outlet-router', component: OutletRouterComponent, data: {pageTitle: 'Allows controlling the web content to be displayed in a router outlet'}},
-  {path: 'publish-message', component: PublishMessageComponent, data: {pageTitle: 'Allows publishing messages'}},
-  {path: 'receive-message', component: ReceiveMessageComponent, data: {pageTitle: 'Allows receiving messages'}},
-  {path: 'register-capability', component: RegisterCapabilityComponent, data: {pageTitle: 'Allows managing capabilities'}},
-  {path: 'lookup-capability', component: LookupCapabilityComponent, data: {pageTitle: 'Allows looking up capabilities'}},
-  {path: 'lookup-intention', component: LookupIntentionComponent, data: {pageTitle: 'Allows looking up intentions'}},
-  {path: 'register-intention', component: RegisterIntentionComponent, data: {pageTitle: 'Allows managing intentions'}},
-  {path: 'context', component: ContextComponent, data: {pageTitle: 'Allows showing the context at this level in the context tree'}},
-  {path: 'microfrontend-1', component: MicrofrontendComponent, data: {pageTitle: 'Displays the \'microfrontend-1\' page'}},
-  {path: 'microfrontend-2', component: MicrofrontendComponent, data: {pageTitle: 'Displays the \'microfrontend-2\' page'}},
-  {path: 'microfrontend-2/:param1/:param2', component: MicrofrontendComponent},
-  {path: 'scrollable-microfrontend', component: ScrollableMicrofrontendComponent, data: {pageTitle: 'Displays a microfrontend with some tall content displayed in a viewport'}},
-  {path: 'preferred-size', component: PreferredSizeComponent, data: {pageTitle: 'Allows playing around with the microfrontend\'s preferred size'}},
-  {path: 'platform-properties', component: PlatformPropertiesComponent, data: {pageTitle: 'Shows properties that are registered in the platform'}},
   {path: 'activator', loadChildren: (): any => import('./activator.module').then(m => m.ActivatorModule)},
   {path: '', redirectTo: 'browser-outlets;count=2', pathMatch: 'full'},
+  {
+    path: '',
+    component: AppShellComponent,
+    children: [
+      {path: 'browser-outlets', component: BrowserOutletsComponent, data: {pageTitle: 'Allows displaying web content in one or more browser outlets', matrixParams: new Map().set('count', 2), pageTitleVisible: false}},
+      {path: 'router-outlet', component: RouterOutletComponent, data: {pageTitle: 'Allows displaying web content in a router outlet'}},
+      {path: 'outlet-router', component: OutletRouterComponent, data: {pageTitle: 'Allows controlling the web content to be displayed in a router outlet'}},
+      {path: 'publish-message', component: PublishMessageComponent, data: {pageTitle: 'Allows publishing messages'}},
+      {path: 'receive-message', component: ReceiveMessageComponent, data: {pageTitle: 'Allows receiving messages'}},
+      {path: 'register-capability', component: RegisterCapabilityComponent, data: {pageTitle: 'Allows managing capabilities'}},
+      {path: 'lookup-capability', component: LookupCapabilityComponent, data: {pageTitle: 'Allows looking up capabilities'}},
+      {path: 'lookup-intention', component: LookupIntentionComponent, data: {pageTitle: 'Allows looking up intentions'}},
+      {path: 'register-intention', component: RegisterIntentionComponent, data: {pageTitle: 'Allows managing intentions'}},
+      {path: 'context', component: ContextComponent, data: {pageTitle: 'Allows showing the context at this level in the context tree'}},
+      {path: 'microfrontend-1', component: MicrofrontendComponent, data: {pageTitle: 'Displays the \'microfrontend-1\' page'}},
+      {path: 'microfrontend-2', component: MicrofrontendComponent, data: {pageTitle: 'Displays the \'microfrontend-2\' page'}},
+      {path: 'microfrontend-2/:param1/:param2', component: MicrofrontendComponent},
+      {path: 'scrollable-microfrontend', component: ScrollableMicrofrontendComponent, data: {pageTitle: 'Displays a microfrontend with some tall content displayed in a viewport'}},
+      {path: 'preferred-size', component: PreferredSizeComponent, data: {pageTitle: 'Allows playing around with the microfrontend\'s preferred size'}},
+      {path: 'platform-properties', component: PlatformPropertiesComponent, data: {pageTitle: 'Shows properties that are registered in the platform'}},
+    ],
+  },
 ];
 
 @NgModule({
