+++ apps/microfrontend-platform-testing-app/src/app/app-shell/app-shell.component.scss
